fixed loading v2
