--- server/.env.example
@@ -1,5 +1,0 @@
-PORT=
-
-HOSTNAME=
-
-DATABASE_URL=
